+++ src/features/App/components/Gw2ApiHealth/index.js
@@ -15,7 +15,7 @@ export default class Gw2ApiHealth extends Component {
 
   componentWillMount () {
     // eslint-disable-next-line
-    axios.get(`${config.gw2.endpoint}/v2/characters?access_token=EE920D9D-F7CF-A146-A5F5-95455980577B0DC68745-969C-4ED9-8462-1299FE6FB078`, { ignoreAuth: true })
+    axios.get(`${config.gw2.endpoint}v2/characters?access_token=EE920D9D-F7CF-A146-A5F5-95455980577B0DC68745-969C-4ED9-8462-1299FE6FB078`, { ignoreAuth: true })
       .then(null, () => {
         this.setState({
           up: false,